:sparkles: 메시지 동기화 서비스 리팩토링
